backend/requirements.txt
asgiref==3.9.1
certifi==2025.10.5
charset-normalizer==3.4.4
Django==4.2.28
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework-simplejwt==5.5.1
idna==3.11
pillow==12.0.0
PyJWT==2.11.0
requests==2.32.5
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
dj-database-url==3.1.0
psycopg2-binary==2.9.11
mongoengine==0.29.1
gunicorn==21.2.0
python-dotenv==1.1.1
whitenoise==6.9.0
cloudinary==1.44.0
django-cloudinary-storage==0.3.0
python-decouple==3.8